employer/api/serializers.py
from rest_framework import serializers
from employer.models import Employee


class EmployeeSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Employee
        fields = "__all__"