requirements.txt
Django==4.2.5
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==23.3
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.4
idna==3.4
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0

simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.5

